rename hash to hashable
