+++ packages/contract-utils/crypto/src/hashable.rs
