+++ packages/contract-utils/crypto/src/keccak.rs
@@ -1,6 +1,6 @@
 use soroban_sdk::{Bytes, BytesN, Env};
 
-use crate::hash::{BuildHasher, Hashable, Hasher};
+use crate::hashable::{BuildHasher, Hashable, Hasher};
 
 pub struct KeccakBuilder {
     env: Env,
