+++ packages/contract-utils/crypto/src/lib.rs
@@ -1,5 +1,5 @@
 #![no_std]
 
-pub mod hash;
+pub mod hashable;
 pub mod keccak;
 pub mod merkle;
